app/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==24.1
phonenumbers==8.13.39
psycopg2-binary==2.9.9
sqlparse==0.5.0
tzdata==2024.1
